Time: 535 ms (5.01%), Space: 117.5 MB (7.37%) - LeetHub

+++ 416-partition-equal-subset-sum/416-partition-equal-subset-sum.java
@@ -1,6 +1,6 @@
 class Solution {
     public boolean canPartition(int[] nums) {
-        HashMap<Double, Boolean> hm = new HashMap<>();
+        HashMap<Integer, Boolean> hm = new HashMap<>();
 
         int sum = 0;
         for(int i=0; i<nums.length ; i++){
@@ -10,7 +10,7 @@ class Solution {
         return targetSum(nums, 0, sum, sum, hm);
     }
     
-   public static boolean targetSum(int[] arr, int idx, double curr, double sum, HashMap<Double, Boolean> hm){
+   public static boolean targetSum(int[] arr, int idx, int curr, int sum, HashMap<Integer, Boolean> hm){
         if(curr == sum/2) return true;
         if(idx>= arr.length || curr<0) return false;
         if(hm.containsKey(curr)) return hm.get(curr);
